requirements.txt
bottle==0.13.4
packaging==25.0
proxy_tools==0.1.0
pycairo==1.29.0
PyGObject==3.50.0
PyGObject-stubs==2.15.0
PyQt6==6.10.1
PyQt6-Qt6==6.10.1
PyQt6-WebEngine==6.10.0
PyQt6-WebEngine-Qt6==6.10.1
PyQt6_sip==13.10.3
pywebview==6.1
QtPy==2.4.3
typing_extensions==4.15.0
